refactored 😌
